Add files via upload
I need to upload my changes to create a pull request, and command line uploads are not working so I'm just going to do it manually - these are all the relevant files I have edited. Again, the main issue I have been having is just properly setting up a filter. The only way I have found to properly reverse the filter is to reverse the filter statement in tags.get but tags.get is called in a way that I don't know how to edit at all